move annotation module from io to util sub-package
