+++ src/vak/io/__init__.py
@@ -1,6 +1,5 @@
 """module that handles file input-output:
 - audio files
 - spectrograms made from audio files of vocalizations
-- associated annotation files
 - .csv files that represent a dataset of vocalizations that combines all those files together"""
-from . import annotation, audio, dataframe, spect
+from . import audio, dataframe, spect
